Correct deal-start detection
In filecoin the payment of a deal can start way later than the actual
sector containing it. This makes the deal retrievable but the accrual of
payment for it is delayed. This is usually done to allow ample time for
sealing to take place, as the start of proving of a containing sector
must be strictly before the deal start.

Within the entire codebase we properly go off dealInfo.State.SectorStartEpoch
except for one forgotten spot, which this commit corrects.

This has very minor implications for the results of the last round of the
competition, inflating the winning numbers by 0.7% ( that's 0.0071 ).
No effort will be taken to correct/exclude any of the deals retroactively.

Going forward the stats should be entirely correct.